fpl-datter-4 add team name and position name into player
